% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotCF.R
\name{barplotCF}
\alias{barplotCF}
\title{Stacked barplot of cell fractions for samples in different groups.}
\usage{
barplotCF(mat, groupInfo = NULL, ctCol = NULL)
}
\arguments{
\item{mat}{a numeric matrix of cell fractions for bulk samples, with sample identifiers as rownames and cell type names as colnames.}

\item{groupInfo}{a vector or factor giving the group names for samples in "\code{mat}". The order of "\code{groupInfo}" should be same as the sample order in "\code{mat}" unless it was named by the sample identifiers. You can designate the bar orders of groups by setting "\code{groupInfo}" to factor format following your interested order. Missing values (NA) will be removed before plotting.}

\item{ctCol}{a character vector specifying the colors of the different cell types, which should be given in the order of the colnames of "\code{mat}". Default colors will be used if not provided.}
}
\description{
Visualize the estimated cell fractions for multiple groups of samples by a stacked bar chart.
}
\examples{
promat <- runif(10 * 7,min = 0, max = 1)
promat <- matrix(promat, nrow = 10)
promat <- promat / rowSums(promat)
rownames(promat) <- paste0("sample_", 1:10)
colnames(promat) <- paste0("ct_", 1:7)
groupinfo <- sample(paste0('Group_', letters[1:3]), 10, replace = T)
barplotCF(promat, groupInfo = groupinfo)
}
