% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotCF.R
\name{boxplotCF}
\alias{boxplotCF}
\title{Boxplot of cell proportions for samples in different groups.}
\usage{
boxplotCF(mat, groupInfo, groupCol = NULL, groupOrder = NULL, use.wilcoxon = T)
}
\arguments{
\item{mat}{see the same argument of barplotCF.}

\item{groupInfo}{see the same argument of barplotCF.}

\item{groupCol}{a character vector specifying the colors of the different group.
Default colors will be used if not provided.}

\item{groupOrder}{a character vector specifying the order of the group.}

\item{use.wilcoxon}{logical, if TRUE, two-sided Wilcoxon rank-sum
tests will be used to assess significance of cell fractions difference
between groups, if FALSE, two-sided t tests will be used. Default is TRUE.
Ignored if there are three or more levels of groupInfo.}
}
\description{
Visualize the estimated immune cell proportions (from the
output of NUWAeDeconv) for multiple groups of samples by boxplots.
}
\examples{
promat <- runif(10 * 7,min = 0, max = 1)
promat <- matrix(promat, nrow = 10)
rownames(promat) <- paste0("sample_", 1:10)
colnames(promat) <- paste0("ct_", 1:7)
promat <- promat / rowSums(promat)
groupinfo <- sample(paste0('Group_', letters[1:3]), 10, replace = T)
boxplotCF(promat, groupInfo = groupinfo)
}
