% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildNetwork}
\alias{buildNetwork}
\title{Build co-expression network for individual marker based on provided training datasets.}
\usage{
buildNetwork(
  trainsets = NULL,
  markers = NULL,
  prep = F,
  batchInfoList = NULL,
  nTr = 2,
  corCutoff = 0.3,
  ncores = 16
)
}
\arguments{
\item{trainsets}{a list containing the datasets used to build the network and
train regression models later. Each dataset of the list should be an numeric
expression matrix with HUGO gene symbols as rownames and sample identifiers
as colnames. Data should be non-logarithm scale. Default is NULL, then a list
including proteomic expression matrices of the six CPTAC datasets
(BRCA, CCRCC, COAD, EC, GC and LUAD) will be used.}

\item{markers}{interesting markers: a character vector of markers, which are
the candidate cores of the network. If NULL (default), the union of markers
from signature matrices "LM6", "LM22"and "BCIC" will be used.}

\item{batchInfoList}{a list containing the batch information corresponding
to the training datasets. the batchInfoList should be named using the names
of trainsets, or the length of batchInfoList should be equal to the number of
trainsets. Each element in the list gives the batch information of one trainset,
which is a vector named by the sample identifiers of that trainset.}

\item{nTr}{a positive integer, we only build network for markers existing
in greater than or equal to nTr training datasets. Default is 3.}

\item{corCutoff}{a positive numeric, specifying the absolute Pearson
correlation coeffienct threshould above which a co-expression will be
declared between individual marker and other quantified protein. For
each marker, we identify its "coherently correlated proteins", i.e.
proteins with the same correlation coeffienct sign in all training
datasets, and with significant correlation (P < 0.05, absolute value
of Pearson correlation coeffienct greater than corCutoff) in at least
two trainign datasets. Default is 0.3.}
}
\value{
A list containing:\describe{
 \item{\code{corr}}{a numeric data frame of Pearson correlation coeffienct between markers and other proteins.}
 \item{\code{markers}}{a character vector, the markers with co-expression networks.}
 \item{\code{trainScaled}}{A list containing the preprocessed training datasets.}
 }
}
\description{
This function builds individual co-expression network for each provided marker,
to select proteins with correlated expression relationship of a marker
(using samples with marker quantification) in the given training datasets
(two or more proteome datasets needed).
}
\examples{
my.net <- buildNetwork(cptacDatasets[1:3])
str(my.net)
}
