% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUWA.mcpcounter.R
\name{NUWA.mcpcounter}
\alias{NUWA.mcpcounter}
\title{A portal function running MCPcounter after NUWAms analysis.}
\usage{
NUWA.mcpcounter(expr, marker_list = NULL, ...)
}
\arguments{
\item{expr}{see the same argument in NUWA.cibersort.}

\item{marker_list}{a list, which names are cellular populations' names and elements are character vectors of markers (HUGO symbols). Default is MCPcounter markers.}

\item{...}{additional arguments passed to the NUWAms() function}
}
\value{
see NUWA.cibersort.
}
\description{
Run NUWAms and MCPcounter algorithm with a marker list of interest.
}
\examples{
expr <- CPTAC.6datasets$brca[, 1:5]
res_nuwa <- NUWA.mcpcounter(expr, marker_list = NULL)
res_nuwa <- NUWA.mcpcounter(expr, marker_list = my_markers)
}
