% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUWA.xcell.R
\name{NUWA.xcell}
\alias{NUWA.xcell}
\title{A portal function running xCell after NUWAms analysis.}
\usage{
NUWA.xcell(expr, marker_list = NULL, ...)
}
\arguments{
\item{expr}{see the same argument in NUWA.cibersort.}

\item{marker_list}{a list, A list where the names represent cellular populations and the elements are character vectors of markers (HUGO symbols). By default, xCell64 markers are used.}

\item{...}{additional arguments passed to the NUWAms() function}
}
\value{
see NUWA.cibersort.
}
\description{
Run NUWAms and xCell algorithm with a marker list of interest.
}
\examples{
expr <- CPTAC.6datasets$brca[, 1:5]
res_nuwa <- NUWA.xcell(expr, marker_list = NULL)
res_nuwa <- NUWA.xcell(expr, marker_list = my_markers)
}
