% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUWAeDeconv.R
\name{NUWAeDeconv}
\alias{NUWAeDeconv}
\title{Immune cell types deconvolution using expression dataset.}
\usage{
NUWAeDeconv(
  m.exp,
  cibersortPath,
  BCIC_min_marker_num = 6,
  LM6_min_marker_num = 6,
  LM22_min_marker_num = 6,
  RNAseq = F,
  protein = F
)
}
\arguments{
\item{m.exp}{a numeric matrix or data frame of expression profiles for bulk
tissue samples, with HUGO gene symbols as rownames and sample identifiers
as colnames. It can also be a string specifing the file path of an expression
matrix. Data must be non-logarithm scale.}

\item{cibersortPath}{a string specifying the path of CIBERSORT R script,
CIBERSORT is only freely available for academic users, please register
on https://cibersort.stanford.edu, and download the CIBERSORT source script.}

\item{BCIC_min_marker_num}{a positive interger, indicating the minimal number
of BCIC markers needed to run EPIC. Default is 6.}

\item{LM6_min_marker_num}{a positive interger, indicating the minimal number
of LM6 markers needed to run CIBERSORT. Default is 6.}

\item{LM22_min_marker_num}{a positive interger, indicating the minimal number
of LM22 markers needed to run CIBERSORT. Default is 6.}

\item{RNAseq}{logical, indciating whether quantile normalization will  be
performed in CIBERSORT analysis. Only set FALSE for RNA-seq data as recommended
on the CIBERSORT website. Default is TRUE.}

\item{protein}{logical, set TRUE when expression matrix is about proteome data.
If TRUE, signature matrix including 118 markers (union of BCIC and TIC markers)
will be used for EPIC analysis, while  the BCIC markers (n = 65) will be used
if FALSE.}
}
\value{
A list:\describe{
 \item{\code{prop}}{ a matrix, the first column is the cell type name, and the
 remaining columns (one sample per column) are the proportion of mRNA or protein
 coming from the six immune cell types (B, CD4 T, CD8 T, monocyte/macrophage,
 NK and neutrophils cells).}
 \item{\code{mergedProp}}{a list containing three merged and one-to-sum
 normalized proportion matrices predicted by CIBERSORT-LM22, CIBERSORT-LM6
 and EPIC-BCIC.}
 \item{\code{rawRes}}{a list containing the raw proportion matrices
 generated by CIBERSORT-LM22, CIBERSORT-LM6 and EPIC-BCIC.}
 \item{\code{usedComb}}{a string vector showing the combinations (from
 CIBERSORT-LM22, CIBERSORT-LM6 and EPIC-BCIC) used to generate the ensembled
 prediction of proportions. Disqualification might be caused by insufficient markers.}
}
}
\description{
This function integrates deconvolution results of three agrithom-signature
combinations selected from our benchmark analysis, and provides the relative
proportions of six immune cell types in mixture samples.
}
\examples{
# You need to provide path to CIBERSORT.R
path='D:/Users/xiergo/Documents/CIBERSORT.R'
res=NUWAeDeconv(m.exp,cibersortPath=path)
}
