% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall.plot.R
\name{recall.plot}
\alias{recall.plot}
\title{Plotting the recall values at both marker level and sample level.}
\usage{
recall.plot(recallRes, level = c("marker", "sample")[1], ...)
}
\arguments{
\item{recallRes}{a list, the output of recall function.}

\item{level}{a character, one of "marker" (at marker level) and "sample" (at sample level). Default is "marker".
At marker level, scatter plot showing associations between marker level recall and correlation coefficients for all samples.
Dotted lines indicate a recall of 0.8, i.e. 80th percentile.
At sample level, density plot showing distributions of correlation coefficients within the same samples or between different samples.
Accuracy rate (AR), representing the overall accuracy in the dataset, and the number of comparisons is indicated.}

\item{...}{additional arguments passed to the ggplot2::theme() function.}
}
\value{
ggplot object
}
\description{
This function outputs a density plot and a scatter points plot to
visualize NUWA-ms accuracy by evaluating similarity between observed
and inferred abundances for markers with quantification in the inferred dataset.
}
\examples{
recall.plot(recallRes, "marker")
recall.plot(recallRes, "sample")
}
