% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotCF.R
\name{boxplotCF}
\alias{boxplotCF}
\title{Boxplot of cell fractions for samples in different groups.}
\usage{
boxplotCF(
  mat,
  groupInfo,
  groupCol = NULL,
  groupOrder = NULL,
  use.wilcoxon = T,
  use.kruskal = T
)
}
\arguments{
\item{mat}{see the same argument of barplotCF.}

\item{groupInfo}{see the same argument of barplotCF.}

\item{groupCol}{a character vector specifying the colors of the different groups. Default colors will be used if not provided.}

\item{groupOrder}{a character vector specifying the order of the group.}

\item{use.wilcoxon}{logical, if TRUE, two-sided Wilcoxon rank-sum
tests will be used to assess significance of cell fractions difference
between groups, if FALSE, two-sided t tests will be used. Default is TRUE.
Ignored if there are three or more levels of groupInfo.}

\item{use.kruskal}{logical, if TRUE, Kruskal-Wallis tests will be used to assess significance of cell fractions difference between groups. If FALSE, one-way ANOVA analysis will be used. Default is TRUE. Ignored if there are only two levels of groupInfo.}
}
\description{
Visualize the estimated cell fractions for multiple groups of samples by boxplots.
}
\details{
P values texted above the boxes are calculated by two-sided Wilcoxon rank-sum tests or two-sided t tests between two groups. If there are three or more groups, P values are generated by Kruskal-Wallis tests or one-way ANOVA analysis. P values colored in red indicate significant differences (P < 0.05).
}
\examples{
promat <- runif(10 * 7,min = 0, max = 1)
promat <- matrix(promat, nrow = 10)
rownames(promat) <- paste0("sample_", 1:10)
colnames(promat) <- paste0("ct_", 1:7)
promat <- promat / rowSums(promat)
groupinfo <- sample(paste0('Group_', letters[1:3]), 10, replace = T)
boxplotCF(promat, groupInfo = groupinfo)
}
