% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildNetwork}
\alias{buildNetwork}
\title{Build co-expression network for individual marker based on provided training datasets.}
\usage{
buildNetwork(
  trainsets = NULL,
  markers = NULL,
  preprocess = T,
  batchInfoList = NULL,
  nTr = 2,
  corCutoff = 0.3,
  ncores = 16
)
}
\arguments{
\item{trainsets}{a list containing the underlying datasets used for building the co-expression networks and for training regression models in the subsequent analysis. Each dataset in the list should be a numeric expression matrix (non-logarithm scale), with HUGO gene symbols as rownames and sample identifiers as colnames. When default (NULL) is applied, a list including CPTAC proteomic datasets of six cancer types (breast cancer, clear cell renal cell carcinoma, colon adenocarcinoma, endometrial carcinoma, gastric cancer, lung adenocarcinoma) will be used (CPTAC.6datasets). Users can provide customized datasets with more specific context, e.g., multiple datasets for a specific cancer type.}

\item{markers}{a character vector of interesting proteins (termed as marker) that need to build co-expression networks. If NULL (default) provided, a pre-defined list MARKERS_Immune.1114 (1114 immune marker proteins collected from public studies) will be used. Another two pre-defined lists, including MARKER_FDA_Drug_Targets.812 (FDA-approved drug targets from HPA and Drugbank) and MARKERS_Immune.NUWAp26 (631 immune cell markers from proteomic signature matrix NUWAp26 we developed for 26 immune cell types) were also provided for convenience.}

\item{preprocess}{logical. If TURE, each matrix in "trainsets" is preprocessed before network building. Default is TRUE. See the Methods section of the NUWA manuscript for more details.}

\item{batchInfoList}{a list containing the batch information corresponding to the training datasets. The batchInfoList should be named using the names of trainsets, or the length of batchInfoList should be equal to the number of trainsets. Each element in the list gives the batch information of one trainset, which is a vector named by the sample identifiers of that trainset.}

\item{nTr}{a positive integer, we only build network for markers existing in greater than or equal to "nTr" training datasets. Default is 2.}

\item{corCutoff}{a positive numeric, specifying the absolute Pearson correlation coefficient threshold above which a co-expression will be declared between individual marker and other quantified protein. For each marker, we identify its "coherently correlated proteins", i.e., proteins with the same correlation coefficient sign in all training datasets, and with significant correlation (P < 0.05, absolute value of Pearson correlation coefficient greater than "corCutoff") in at least two training datasets. Default is 0.3.}

\item{ncores}{a positive integer, indicating the number of cores used by this function. If the operating system is windows, then only one core will be used.}
}
\value{
A list containing:\describe{
 \item{\code{corr}}{a numeric data frame of Pearson correlation coefficients between markers and other proteins.}
 \item{\code{markers}}{a character vector, containing the markers with co-expression networks.}
 \item{\code{trainScaled}}{a list containing the preprocessed training datasets.}
 }
}
\description{
This function builds individual co-expression network for each provided marker,
to select proteins with correlated expression relationship of a marker
(using samples with marker quantification) in the given training datasets
(two or more proteome datasets needed).
}
\examples{
my.net <- buildNetwork(CPTAC.6datasets[1:5])
str(my.net)

}
