% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUWA.EPIC.R
\name{NUWA.EPIC}
\alias{NUWA.EPIC}
\title{Built-in NUWA analysis}
\usage{
NUWA.EPIC(expr, signature_matrix)
}
\arguments{
\item{expr}{see the same argument in NUWA.cibersort.}

\item{signature_matrix}{see the same argument in NUWA.cibersort.}
}
\value{
see NUWA.cibersort.
}
\description{
run NUWAms and EPIC algorithm with interested signature matrix
}
\examples{
expr <- cptacDatasets$brca[, 1:5]
res_nuwa <- NUWA.EPIC(expr, signature_matrix = BCIC)
res_nuwa <- NUWA.EPIC(expr, signature_matrix = TIC)
res_nuwa <- NUWA.EPIC(expr, signature_matrix = my_signature_matrix)
}
