% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUWA.mcpcounter.R
\name{NUWA.mcpcounter}
\alias{NUWA.mcpcounter}
\title{A portal function running CIBERSORT after NUWAms analysis.}
\usage{
NUWA.mcpcounter(expr, marker_list = NULL)
}
\arguments{
\item{expr}{see the same argument in NUWA.cibersort.}

\item{marker_list}{see the same argument in NUWA.xcell, default is MCPcounter markers.}
}
\value{
see NUWA.cibersort.
}
\description{
Run NUWAms and MCPcounter algorithm with a marker list of interest.
}
\examples{
expr <- cptacDatasets$brca[, 1:5]
res_nuwa <- NUWA.mcpcounter(expr, marker_list = NULL)
res_nuwa <- NUWA.mcpcounter(expr, marker_list = my_markers)
}
