% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUWA.mcpcounter.R
\name{NUWA.mcpcounter}
\alias{NUWA.mcpcounter}
\title{Built-in NUWA analysis}
\usage{
NUWA.mcpcounter(expr, marker_list = NULL)
}
\arguments{
\item{expr}{see the same argument in NUWA.cibersort.}

\item{marker_list}{see the same argument in NUWA.xcell, default is MCPcounter markers.}
}
\value{
see NUWA.cibersort.
}
\description{
run NUWAms and MCPcounter algorithm with interested marker list.
}
\examples{
res_nuwa <- NUWA.mcpcounter(expr = raw_expr, marker_list = NULL)
res_nuwa <- NUWA.mcpcounter(expr = raw_expr, marker_list = my_markers)
}
