% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUWA.xcell.R
\name{NUWA.xcell}
\alias{NUWA.xcell}
\title{Built-in NUWA analysis}
\usage{
NUWA.xcell(expr, marker_list = NULL)
}
\arguments{
\item{expr}{see the same argument of NUWA.cibersort.}

\item{marker_list}{a list, whose names are cellular populations' names
and elements are character vectors of markers (HUGO symbols), default is
xCell64.}
}
\value{
see NUWA.cibersort.
}
\description{
run NUWAms and xCell algorithm with interested marker list.
}
\examples{
res_nuwa <- NUWA.xcell(expr = raw_expr, marker_list = NULL)
res_nuwa <- NUWA.xcell(expr = raw_expr, marker_list = my_markers)
}
