% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUWAeDeconv.R
\name{NUWAeDeconv}
\alias{NUWAeDeconv}
\title{Immune cell types deconvolution using expression dataset.}
\usage{
NUWAeDeconv(
  expr,
  cibersortPath,
  BCIC_min_marker_num = 6,
  LM6_min_marker_num = 6,
  LM22_min_marker_num = 6,
  quantileNorm_cibersort = T,
  protein = T
)
}
\arguments{
\item{expr}{a numeric matrix of expression profiles for bulk tissue samples, with HUGO gene symbols as rownames and sample identifiers as colnames. Data must be non-logarithm scale.}

\item{cibersortPath}{a string specifying the path of CIBERSORT R script, CIBERSORT is only freely available for academic users, please register on \url{https://cibersort.stanford.edu}, and download the CIBERSORT source script.}

\item{BCIC_min_marker_num}{a positive integer, indicating the minimal number of BCIC markers needed to run EPIC. Default is 6.}

\item{LM6_min_marker_num}{a positive integer, indicating the minimal number of LM6 markers needed to run CIBERSORT. Default is 6.}

\item{LM22_min_marker_num}{a positive integer, indicating the minimal number of LM22 markers needed to run CIBERSORT. Default is 6.}

\item{quantileNorm_cibersort}{logical, indicating whether quantile normalization will be performed in CIBERSORT analysis. Only set FALSE for RNA-seq data as recommended on the CIBERSORT website. Default is TRUE.}

\item{protein}{logical, set TRUE for proteomic expression data. If TRUE, signature matrix including 118 markers (union of BCIC and TIC markers) will be used for EPIC analysis. If FALSE, the BCIC markers (n = 65) will be used. Default is TRUE.}
}
\value{
A list containing:\describe{
 \item{\code{proportion}}{ a matrix, the first column is the cell type name, and the remaining columns (one sample per column) are the relative proportion of mRNA or protein coming from the six immune cell types (B, CD4 T, CD8 T, monocyte/macrophage, NK and neutrophils cells).}
 \item{\code{mergedProp}}{a list containing three merged and sum-to-one normalized proportion matrices predicted by CIBERSORT-LM22, CIBERSORT-LM6 and EPIC-BCIC.}
 \item{\code{rawRes}}{a list containing the raw proportion matrices generated by CIBERSORT-LM22, CIBERSORT-LM6 and EPIC-BCIC.}
 \item{\code{usedComb}}{a string vector showing the combinations (from CIBERSORT-LM22, CIBERSORT-LM6 and EPIC-BCIC) used to generate the ensembled prediction of proportions. Disqualification might be caused by insufficient markers.}
}
}
\description{
This function integrates deconvolution results of three algorithm-signature combinations selected from our benchmark analysis, and provides the relative proportions of six immune cell types in mixture samples. See the NUWA manuscript for more details.
}
\examples{
# You need to provide path to CIBERSORT.R
# cibersortPath = "<PATHTO>/CIBERSORT.R"
my.nuwadec = NUWAeDeconv(expr=my.nuwams$finalExpr, cibersortPath= cibersortPath)
}
